rag/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
distro==1.9.0
    # via openai
faiss-cpu==1.11.0
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via httpcore
hf-xet==1.1.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
huggingface-hub==0.33.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via torch
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via -r requirements.in
langchain-core==0.3.65
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.45
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.0
    # via
    #   faiss-cpu
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
openai==1.86.0
    # via -r requirements.in
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   transformers
pandas==2.3.0
    # via -r requirements.in
pillow==11.2.1
    # via sentence-transformers
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   transformers
regex==2024.11.6
    # via
    #   tiktoken
    #   transformers
requests==2.32.4
    # via
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via sentence-transformers
scipy==1.15.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.1.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via langchain
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via -r requirements.in
tokenizers==0.21.1
    # via transformers
torch==2.7.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.52.4
    # via sentence-transformers
typing-extensions==4.14.0
    # via
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
zstandard==0.23.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# setuptools
